config/docker/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai>=1.0.0
apscheduler==3.10.4
requests==2.31.0
httpx==0.25.2
